// src/lib/types.ts
// Make enum nullable

enum Degree {
  DTEK = "DTEK",
  DSIK = "DSIK",
  DVIT = "DVIT",
  BINF = "BINF",
  IMO = "IMO",
  // IKT and KOGNI should not be used,
  // are only here for backwards compatibility.
  IKT = "IKT",
  KOGNI = "KOGNI",
  //
  INF = "INF",
  PROG = "PROG",
  ARMNINF = "ARMNINF",
  POST = "POST",
  MISC = "MISC",
}

export { Degree };